Added library for PDF generation
- Added file read and write permission request to save PDF documents;
- Refactored value type for quantity of products, now accepting Float type;
- Added method to generate a PDF receipt with the purchase summary;
- Configured file creation.

+++ app/src/main/java/com/rodrigobn/frameworkdigitale_commerce/data/models/Product.kt
@@ -14,7 +14,7 @@ class Product(
     @ColumnInfo
     var description: String,
     @ColumnInfo
-    var quantity: Int,
+    var quantity: Float,
     @ColumnInfo
     var price: Float,
     @ColumnInfo
